+++ mne/minimum_norm/tests/test_inverse.py
@@ -14,12 +14,14 @@ from mne.source_estimate import read_source_estimate
 from mne import fiff, read_cov, read_forward_solution
 from mne.minimum_norm.inverse import apply_inverse, read_inverse_operator, \
     apply_inverse_raw, apply_inverse_epochs, make_inverse_operator, \
-    write_inverse_operator
+    write_inverse_operator, compute_rank_inverse
 from mne.utils import _TempDir
 
 s_path = op.join(sample.data_path(), 'MEG', 'sample')
 fname_inv = op.join(s_path, 'sample_audvis-meg-oct-6-meg-inv.fif')
 fname_inv_fixed = op.join(s_path, 'sample_audvis-meg-oct-6-meg-fixed-inv.fif')
+fname_inv_diag = op.join(s_path,
+                         'sample_audvis-meg-oct-6-meg-diagnoise-inv.fif')
 fname_vol_inv = op.join(s_path, 'sample_audvis-meg-vol-7-meg-inv.fif')
 fname_data = op.join(s_path, 'sample_audvis-ave.fif')
 fname_cov = op.join(s_path, 'sample_audvis-cov.fif')
@@ -33,6 +35,8 @@ label_lh = read_label(fname_label % 'Aud-lh')
 label_rh = read_label(fname_label % 'Aud-rh')
 noise_cov = read_cov(fname_cov)
 raw = fiff.Raw(fname_raw)
+evoked = fiff.Evoked(fname_data, setno=0, baseline=(None, 0))
+evoked.crop(0, 0.2)
 snr = 3.0
 lambda2 = 1.0 / snr ** 2
 
@@ -40,12 +44,18 @@ tempdir = _TempDir()
 
 
 def _compare(a, b):
+    skip_types = ['whitener', 'proj', 'reginv', 'noisenorm', 'nchan']
     if isinstance(a, dict):
         assert_true(isinstance(b, dict))
         for k, v in a.iteritems():
-            if not k in b:
-                raise ValueError('%s not in %s' % (k, b))
+            if not k in b and k not in skip_types:
+                raise ValueError('First one had one setond one didnt:\n'
+                                 '%s not in %s' % (k, b.keys()))
             _compare(v, b[k])
+        for k, v in b.iteritems():
+            if not k in a and k not in skip_types:
+                raise ValueError('Second one had one first one didnt:\n'
+                                 '%s not in %s' % (k, a.keys()))
     elif isinstance(a, list):
         assert_true(len(a) == len(b))
         for i, j in zip(a, b):
@@ -60,33 +70,59 @@ def _compare(a, b):
         assert_true(a == b)
 
 
-def test_io_inverse_operator():
-    """Test IO of inverse_operator
-    """
-    inverse_operator_vol = read_inverse_operator(fname_vol_inv)
-    for inv in [inverse_operator, inverse_operator_vol]:
-        inv_init = copy.deepcopy(inv)
-        out_file = op.join(tempdir, 'test-inv.fif')
-        write_inverse_operator(out_file, inv)
-        this_inv = read_inverse_operator(out_file)
-        _compare(inv, inv_init)
-        _compare(inv, this_inv)
+def _compare_inverses_approx(inv_1, inv_2, evoked, stc_decimals):
+    if inv_1['depth_prior'] is not None:
+        assert_array_almost_equal(inv_1['depth_prior']['data'],
+                                  inv_2['depth_prior']['data'])
+    else:
+        assert_true(inv_2['depth_prior'] is None)
+    if inv_1['orient_prior'] is not None:
+        assert_array_almost_equal(inv_1['orient_prior']['data'],
+                                  inv_2['orient_prior']['data'])
+    else:
+        assert_true(inv_2['orient_prior'] is None)
+    assert_array_almost_equal(inv_1['source_cov']['data'],
+                              inv_2['source_cov']['data'])
 
-    # just do one example for .gz, as it should generalize
-    inv = inverse_operator
-    out_file = op.join(tempdir, 'test-inv.fif.gz')
-    write_inverse_operator(out_file, inv)
-    this_inv = read_inverse_operator(out_file)
-    _compare(inv, this_inv)
+    #assert_array_almost_equal(inv_1['eigen_fields']['data'],
+    #                          inv_2['eigen_fields']['data'])
+    #assert_array_almost_equal(inv_1['eigen_leads']['data'],
+    #                          inv_2['eigen_leads']['data'])
+
+    stc_1 = apply_inverse(evoked, inv_1, lambda2, "dSPM")
+    stc_2 = apply_inverse(evoked, inv_2, lambda2, "dSPM")
+
+    assert_equal(stc_1.times, stc_2.times)
+    assert_array_almost_equal(stc_1.data, stc_2.data, stc_decimals)
+
+
+def _compare_io(inv_op, out_file_ext='.fif'):
+    if out_file_ext == '.fif':
+        out_file = op.join(tempdir, 'test-inv.fif')
+    elif out_file_ext == '.gz':
+        out_file = op.join(tempdir, 'test-inv.fif.gz')
+    else:
+        raise ValueError('IO test could not complete')
+    # Test io operations
+    inv_init = copy.deepcopy(inv_op)
+    write_inverse_operator(out_file, inv_op)
+    read_inv_op = read_inverse_operator(out_file)
+    _compare(inv_init, read_inv_op)
+    _compare(inv_init, inv_op)
 
 
 def test_apply_inverse_operator():
-    """Test MNE inverse computation
-
-    With and without precomputed inverse operator.
+    """Test MNE inverse computation (precomputed and non-precomputed)
     """
-    evoked = fiff.Evoked(fname_data, setno=0, baseline=(None, 0))
-    evoked.crop(0, 0.1)
+
+    # Test MNE inverse computation starting from forward operator
+    fwd_op = read_forward_solution(fname_fwd, surf_ori=True)
+    my_inv_op = make_inverse_operator(evoked.info, fwd_op, noise_cov,
+                                      loose=0.2, depth=0.8)
+    _compare_io(my_inv_op)
+    _compare_inverses_approx(my_inv_op, inverse_operator, evoked, 2)
+    # Inverse has 306 channels - 4 proj = 302
+    assert_true(compute_rank_inverse(inverse_operator) == 302)
 
     stc = apply_inverse(evoked, inverse_operator, lambda2, "MNE")
     assert_true(stc.data.min() > 0)
@@ -103,16 +139,6 @@ def test_apply_inverse_operator():
     assert_true(stc.data.max() < 35)
     assert_true(stc.data.mean() > 0.1)
 
-    # Test MNE inverse computation starting from forward operator
-    fwd_op = read_forward_solution(fname_fwd, surf_ori=True)
-
-    my_inv_op = make_inverse_operator(evoked.info, fwd_op, noise_cov,
-                                      loose=0.2, depth=0.8)
-    out_file = op.join(tempdir, 'test-inv.fif')
-    write_inverse_operator(out_file, my_inv_op)
-    read_my_inv_op = read_inverse_operator(out_file)
-    _compare(my_inv_op, read_my_inv_op)
-
     my_stc = apply_inverse(evoked, my_inv_op, lambda2, "dSPM")
 
     assert_true('dev_head_t' in my_inv_op['info'])
@@ -123,39 +149,38 @@ def test_apply_inverse_operator():
 
 
 def test_make_inverse_operator_fixed():
-    """Test MNE inverse computation with fixed orientation"""
-    # XXX : should be fixed and not skipped
-    raise nose.SkipTest("XFailed Test")
-
-    evoked = fiff.Evoked(fname_data, setno=0, baseline=(None, 0))
+    """Test MNE inverse computation with fixed orientation & no depth weighting
+    """
     fwd_op = read_forward_solution(fname_fwd, force_fixed=True)
-    inv_op = make_inverse_operator(evoked.info, fwd_op, noise_cov, depth=0.8,
+    inv_op = make_inverse_operator(evoked.info, fwd_op, noise_cov, depth=None,
                                    loose=None)
+    _compare_io(inv_op)
     inverse_operator_fixed = read_inverse_operator(fname_inv_fixed)
+    # XXX The STCs are not that equivalent in the fixed case...
+    _compare_inverses_approx(inverse_operator_fixed, inv_op, evoked, -1)
+    # Inverse has 306 channels - 4 proj = 302
+    assert_true(compute_rank_inverse(inverse_operator_fixed) == 302)
 
-    assert_array_almost_equal(inverse_operator_fixed['depth_prior']['data'],
-                              inv_op['depth_prior']['data'])
-    assert_equal(inverse_operator_fixed['orient_prior'],
-                 inv_op['orient_prior'])
-    assert_array_almost_equal(inverse_operator_fixed['source_cov']['data'],
-                              inv_op['source_cov']['data'])
 
-    stc_fixed = apply_inverse(evoked, inverse_operator_fixed, lambda2, "dSPM")
-    my_stc = apply_inverse(evoked, inv_op, lambda2, "dSPM")
-
-    assert_equal(stc_fixed.times, my_stc.times)
-    assert_array_almost_equal(stc_fixed.data, my_stc.data, 2)
-
-    # assert_array_almost_equal(inverse_operator_fixed['eigen_fields']['data'],
-    #                           inv_op['eigen_fields']['data'])
-    # assert_array_almost_equal(inverse_operator_fixed['eigen_leads']['data'],
-    #                           inv_op['eigen_leads']['data'])
+def test_make_inverse_operator_diag():
+    """Test MNE inverse computation with diagonal noise cov
+    """
+    fwd_op = read_forward_solution(fname_fwd, surf_ori=True)
+    inv_op = make_inverse_operator(evoked.info, fwd_op, noise_cov.as_diag(),
+                                   loose=0.2, depth=0.8)
+    _compare_io(inv_op)
+    inverse_operator_diag = read_inverse_operator(fname_inv_diag)
+    # XXX This one's only good to zero decimal places
+    _compare_inverses_approx(inverse_operator_diag, inv_op, evoked, 0)
+    # Inverse has 306 channels - 4 proj = 302
+    assert_true(compute_rank_inverse(inverse_operator_diag) == 302)
 
 
 def test_inverse_operator_volume():
-    """Test MNE inverse computation on volume source space"""
-    evoked = fiff.Evoked(fname_data, setno=0, baseline=(None, 0))
+    """Test MNE inverse computation on volume source space
+    """
     inverse_operator_vol = read_inverse_operator(fname_vol_inv)
+    _compare_io(inverse_operator_vol)
     stc = apply_inverse(evoked, inverse_operator_vol, lambda2, "dSPM")
     stc.save(op.join(tempdir, 'tmp-vl.stc'))
     stc2 = read_source_estimate(op.join(tempdir, 'tmp-vl.stc'))
@@ -165,8 +190,16 @@ def test_inverse_operator_volume():
     assert_array_almost_equal(stc.times, stc2.times)
 
 
+def test_io_inverse_operator():
+    """Test IO of inverse_operator with GZip
+    """
+    # just do one example for .gz, as it should generalize
+    _compare_io(inverse_operator, '.gz')
+
+
 def test_apply_mne_inverse_raw():
-    """Test MNE with precomputed inverse operator on Raw"""
+    """Test MNE with precomputed inverse operator on Raw
+    """
     start = 3
     stop = 10
     _, times = raw[0, start:stop]
@@ -190,7 +223,8 @@ def test_apply_mne_inverse_raw():
 
 
 def test_apply_mne_inverse_fixed_raw():
-    """Test MNE with fixed-orientation inverse operator on Raw"""
+    """Test MNE with fixed-orientation inverse operator on Raw
+    """
     start = 3
     stop = 10
     _, times = raw[0, start:stop]