+++ doc/source/whats_new.rst
@@ -11,6 +11,7 @@ Changelog
 
    - Linear (and zeroth-order) detrending for Epochs and Evoked by `Eric Larson`_
 
+   - Improved diagonal noise covariance, depth weighting, and rank computing support by `Eric Larson`_
 
 .. _changes_0_5: